api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "chroma>=0.2.0",
    "chromadb>=1.3.3",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.1",
    "google-generativeai>=0.3.2",
    "grandalf>=0.8",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "huggingface-hub>=0.36.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.1",
    "langchain-google-genai>=0.0.1",
    "langchain-huggingface>=1.0.0",
    "langchain-openai>=1.0.1",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "openai>=2.6.1",
    "pydantic[email]>=2.12.3",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
]
